% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_read.R
\name{ukb_read}
\alias{ukb_read}
\title{Read in the UKB data}
\usage{
ukb_read(fileset, path = ".", ...)
}
\arguments{
\item{fileset}{The prefix for a UKB fileset, e.g., ukbxxxx (for ukbxxxx.tab, ukbxxxx.r, ukbxxxx.html)}

\item{path}{The path to the directory containing your UKB fileset. The default value is the current directory.}
}
\value{
UK Biobank set in a data.frame with usable column names
}
\description{
Read in the UKB data
}
