% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbfetch.R
\name{ukbfetch}
\alias{ukbfetch}
\title{Fetch UK Biobank bulk data files}
\usage{
ukbfetch(participant_ID, data_ID, tools_path = "", auth_file = NULL,
  verbose = FALSE, run_local = TRUE, create_batch = TRUE)
}
\arguments{
\item{participant_ID}{list of subject IDs for which
we want to download bulk imaging files}

\item{data_ID}{list of data IDs to be retrieved}

\item{tools_path}{provide location of the ukbfetch binary
and also `auth_file` / .ukbkey file location}

\item{auth_file}{provide optional key file (in case .ukbkey
is not present)}

\item{verbose}{should the output be verbose?}

\item{run_local}{download the files}

\item{create_batch}{create a batch file to be used
within a terminal}
}
\description{
Fetch UK Biobank bulk data files
}
